requirements.txt
# openface-test==0.1.26  # should be installed separately due to dependency issues with "pip install --no-deps openface-test==0.1.26"
torch==2.9.0; sys_platform != 'win32'  # needed for DL models (On Windows please install from https://pytorch.org/get-started/locally/)
torchvision==0.24.0; sys_platform != 'win32'  # needed for image transformations
numpy==1.26.4
matplotlib==3.10.1
imageio[ffmpeg]==2.34.2
scikit-learn==1.7.2  # needed for traditional ML algorithms
opencv-python==4.11.0.86  # needed for image processing and some out-of-the-box models
ipykernel==7.0.1  # needed for Jupyter notebooks
transformers==4.57.1  # needed for downloading and easy access to pre-trained models
accelerate==1.11.0  # needed for transformers to work properly
furhat-remote-api==1.0.2  # needed for Furhat interaction (Lab 2)
# LLM integration
google-genai==1.46.0  # needed for Google GenAI
# needed by openface
click==8.3.0
pandas==2.3.3
scikit-image==0.25.2
scipy==1.13.1  # old version cause of openface
tensorboardX==2.6.4
timm==1.0.21

langchain==1.1.2
langchain-openai==1.1.0
dotenv==0.9.9
furhat-realtime-api==0.1.3